fix: fixing the bugs and typography in the Volunteer Section and others
