+++ nuestra_senora_informativa/src/Icons/IconsComponent.tsx
@@ -7,19 +7,18 @@ interface IconsComponentProps {
     iconName: string;
 }
 
-const IconMap: { [key: string]: IconType } = {
-  vision: FaBinoculars,
-  mission: FaLightbulb,
-  contact: FaPhoneAlt,
-  email: FaEnvelope,
-  location: FaMapMarkerAlt,
-  networks: FaFacebookF, 
-  about: FaInfoCircle,
+export const iconMap: { [key: string]: IconType } = {
+  "vision": FaBinoculars,
+  "mission": FaLightbulb,
+  "Hogar de Nuestra Señora del Rosario, Santa Cruz, Gte": FaFacebookF,
+  "hogar.nsrosario@yahoo.es": FaEnvelope,
+  "+(506) 2680 0591": FaPhoneAlt,
+  "Santa Cruz, Guanacaste, Costa Rica": FaMapMarkerAlt,
+  "Citas ó visitas a nuestras instalaciones en el horario de lunes a sábados de 8:30 a.m. a 11:00 a.m. y de 1:00 p.m. a 3:30 p.m.": FaInfoCircle
 };
 
-
 const IconsComponent: React.FC<IconsComponentProps> = ({ size, color, iconName }) => {
-  const Icon = IconMap[iconName.toLowerCase()]; // Mapea iconName a su ícono correspondiente
+  const Icon = iconMap[iconName.toLowerCase()]; // Mapea iconName a su ícono correspondiente
 
   if (!Icon) {
     return <span>Icono no encontrado</span>; // Muestra un mensaje si el ícono no existe en IconMap
