+++ nuestra_senora_informativa/src/Services/ServiceInformative.ts
@@ -1,16 +1,9 @@
-
-
-
 import axios from 'axios';
-
-
 
 const URLBase =  'http://localhost:5074/api';
 
-
 export const getAboutUsSection = async () => {
   const response = await axios.get(`${URLBase}/AboutUsSection`);
-  console.log(response.data);
   return response.data;
 }
 
@@ -31,7 +24,6 @@ export const getTitles = async (id: number) => {
 
 export const getIcons = async () => {
   const response = await axios.get(`${URLBase}/SiteSettings`);
- console.log(response.data);
   return response.data;
 };
 
@@ -45,19 +37,7 @@ export const getButtonText = async (id: number) => {
   return response.data;
 };
 
-
-
 export const getHeroSection = async () => {
 const response = await axios.get(`${URLBase}/HeroSection`);
 return response.data ; 
-
 }
-
-
-
-export const SiteSettings = async () => {
-  const response = await axios.get(`${URLBase}/SiteSettings`);
-  return response.data;
-}
-
-
